:hammer: (tooltip): separate container & content classes
