+++ packages/@ourworldindata/grapher/src/tooltip/Tooltip.tsx
@@ -1,22 +1,13 @@
 import React from "react"
 import classnames from "classnames"
-import { CoreColumn } from "@ourworldindata/core-table"
 import { observable, computed, action } from "mobx"
 import { observer } from "mobx-react"
-import { NO_DATA_LABEL } from "../color/ColorScale.js"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons"
-import { Bounds, PointVector, sum, zip } from "@ourworldindata/utils"
-import {
-    TooltipProps,
-    TooltipManager,
-    TooltipTableProps,
-    TooltipValueProps,
-    TooltipValueRangeProps,
-    TooltipFadeMode,
-} from "./TooltipProps"
+import { Bounds, PointVector } from "@ourworldindata/utils"
+import { TooltipProps, TooltipManager, TooltipFadeMode } from "./TooltipProps"
+export * from "./TooltipContents.js"
 
-export const NO_DATA_COLOR = "#999"
 export const TOOLTIP_FADE_DURATION = 400 // $fade-time + $fade-delay in scss
 
 export class TooltipState<T> {
@@ -58,234 +49,6 @@ export class TooltipState<T> {
     get fading(): TooltipFadeMode | undefined {
         // returns "delayed"|"immediate" during the timeout after clearing the target
         return !!this._timer && !!this._target ? this._fade : undefined
-    }
-}
-
-export class TooltipValue extends React.Component<TooltipValueProps> {
-    render(): JSX.Element | null {
-        const { column, value, color, notice } = this.props,
-            displayValue =
-                (value !== undefined && column.formatValueShort(value)) ||
-                NO_DATA_LABEL,
-            displayColor =
-                displayValue === NO_DATA_LABEL ? NO_DATA_COLOR : color
-
-        return (
-            <Indicator column={column} color={displayColor} notice={notice}>
-                {displayValue}
-            </Indicator>
-        )
-    }
-}
-
-export class TooltipValueRange extends React.Component<TooltipValueRangeProps> {
-    ARROW_PATHS = {
-        up: "m14,0H5c-.552,0-1,.448-1,1s.448,1,1,1h6.586L.29303,13.29297l1.41394,1.414L13,3.41394v6.58606c0,.552.448,1,1,1s1-.448,1-1V1c0-.552-.448-1-1-1Z",
-        down: "m14,4c-.552,0-1,.448-1,1v6.586L1.56049.14648.14655,1.56042l11.43958,11.43958h-6.58612c-.552,0-1,.448-1,1s.448,1,1,1h9c.552,0,1-.448,1-1V5c0-.552-.448-1-1-1Z",
-        right: "m19.59198,6.82422L13.22803.46021c-.39105-.39099-1.02405-.39099-1.414,0-.39105.39001-.39105,1.02405,0,1.414l4.65698,4.65704H.5v2h15.97101l-4.65698,4.65698c-.39105.39001-.39105,1.02399,0,1.414.38995.39099,1.02295.39099,1.414,0l6.36395-6.36401c.39001-.39001.39001-1.02399,0-1.414Z",
-    }
-
-    arrowIcon(direction: "up" | "right" | "down"): JSX.Element {
-        return (
-            <svg
-                className={classnames("arrow", direction)}
-                xmlns="http://www.w3.org/2000/svg"
-                viewBox={`0 0 ${direction == "right" ? 20 : 15} 15`}
-            >
-                <path d={this.ARROW_PATHS[direction]} />
-            </svg>
-        )
-    }
-
-    render(): JSX.Element | null {
-        const { column, values, color, notice } = this.props,
-            [firstValue, lastValue] = values.map((v) =>
-                column.formatValueShort(v)
-            ),
-            [firstTerm, lastTerm] =
-                // TODO: would be nicer to actually measure the typeset text
-                sum([firstValue?.length, lastValue?.length]) > 20
-                    ? values.map((v) =>
-                          column.formatValueShortWithAbbreviations(v)
-                      )
-                    : [firstValue, lastValue],
-            trend =
-                values.length < 2
-                    ? null
-                    : values[0] < values[1]
-                    ? this.arrowIcon("up")
-                    : values[0] > values[1]
-                    ? this.arrowIcon("down")
-                    : this.arrowIcon("right")
-
-        return (
-            <Indicator column={column} color={color} notice={notice}>
-                <span className="range">
-                    {firstTerm} {trend} {lastTerm}
-                </span>
-            </Indicator>
-        )
-    }
-}
-
-class Indicator extends React.Component<{
-    column: CoreColumn
-    color?: string
-    notice?: number | string
-    children?: React.ReactNode
-}> {
-    render(): JSX.Element | null {
-        const { column, children, color, notice } = this.props
-        if (column.isMissing || column.name == "time") return null
-
-        const { displayName, unit, shortUnit } = column,
-            displayUnit =
-                unit && displayName != unit && unit != shortUnit
-                    ? unit.replace(/(^\(|\)$)/g, "")
-                    : null,
-            noticeSpan = notice && (
-                <span className="notice">
-                    <FontAwesomeIcon icon={faInfoCircle} />
-                    {notice}
-                </span>
-            )
-        return (
-            <div className="variable">
-                <div className="definition">
-                    <b>{displayName}</b>
-                    {displayUnit && <em> ({displayUnit})</em>}
-                </div>
-                <div className="values" style={{ color }}>
-                    {children}
-                    {noticeSpan}
-                </div>
-            </div>
-        )
-    }
-}
-
-export class TooltipTable extends React.Component<TooltipTableProps> {
-    render(): JSX.Element | null {
-        const { columns, totals, rows } = this.props,
-            focal = rows.some((row) => row.focused),
-            swatched = rows.some((row) => row.swatch),
-            format = { trailingZeroes: true, ...this.props.format },
-            tooEmpty =
-                rows.length < 2 ||
-                totals?.every((value) => value === undefined) ||
-                rows.some(({ values }) =>
-                    values.every((value) => value === undefined)
-                ),
-            tooTrivial = zip(columns, totals ?? []).every(
-                ([column, total]) =>
-                    !!column?.formatValueShort(total).match(/^100(\.0+)?%/)
-            )
-
-        return (
-            <table className={classnames({ focal, swatched })}>
-                {columns.length > 1 && (
-                    <thead>
-                        <tr>
-                            <td className="series-color"></td>
-                            <td className="series-name"></td>
-                            {columns.map((column) => (
-                                <td className="series-value" key={column.slug}>
-                                    {column.displayName}
-                                </td>
-                            ))}
-                        </tr>
-                    </thead>
-                )}
-                <tbody>
-                    {rows.map((row) => {
-                        const {
-                            name,
-                            focused,
-                            blurred,
-                            annotation,
-                            values,
-                            notice,
-                            swatch = "transparent",
-                        } = row
-                        const [_m, seriesName, seriesParenthetical] =
-                            name.match(/^(.*?)(\(.*)?$/) ?? []
-
-                        return (
-                            <tr
-                                key={name}
-                                className={classnames({ focused, blurred })}
-                            >
-                                <td className="series-color">
-                                    <div
-                                        className="swatch"
-                                        style={{ backgroundColor: swatch }}
-                                    />
-                                </td>
-                                <td className="series-name">
-                                    {seriesName}
-                                    {seriesParenthetical && (
-                                        <span className="parenthetical">
-                                            {seriesParenthetical}
-                                        </span>
-                                    )}
-                                    {annotation && (
-                                        <span className="annotation">
-                                            {annotation}
-                                        </span>
-                                    )}
-                                </td>
-                                {zip(columns, values).map(([column, value]) => {
-                                    const missing = value === undefined
-                                    return column ? (
-                                        <td
-                                            key={column.slug}
-                                            className={classnames(
-                                                "series-value",
-                                                { missing }
-                                            )}
-                                        >
-                                            {!missing &&
-                                                column.formatValueShort(
-                                                    value,
-                                                    format
-                                                )}
-                                        </td>
-                                    ) : null
-                                })}
-                                {notice && (
-                                    <td className="notice">
-                                        <FontAwesomeIcon icon={faInfoCircle} />{" "}
-                                        {notice}
-                                    </td>
-                                )}
-                            </tr>
-                        )
-                    })}
-                    {totals && !(tooEmpty || tooTrivial) && (
-                        <>
-                            <tr className="spacer"></tr>
-                            <tr className="total">
-                                <td className="series-color"></td>
-                                <td className="series-name">Total</td>
-                                {zip(columns, totals).map(([column, total]) => (
-                                    <td
-                                        key={column?.slug}
-                                        className="series-value"
-                                    >
-                                        {column && total !== undefined
-                                            ? column.formatValueShort(
-                                                  total,
-                                                  format
-                                              )
-                                            : null}
-                                    </td>
-                                ))}
-                            </tr>
-                        </>
-                    )}
-                </tbody>
-            </table>
-        )
     }
 }
